requirements.txt
pydantic==2.12.5
fastapi==0.127.0
uvicorn==0.40.0
starlette==0.50.0
langgraph==1.0.5
numpy==1.26.4
requests==2.32.5
python-dotenv==1.2.1
torch==2.9.1
torchvision==0.24.1
langchain==1.2.0
langchain-core==1.2.2
langchain_google_genai==4.1.1